backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
openpyxl
pdfplumber
python-dotenv
